Render titles as HTML in ProductName component and use Typescript
